requirements.txt
requests>=2.32.3
python-dotenv>=1.0.1
python-dateutil>=2.8.2
ag2[openai]==0.8.6
pydantic>=2.6.1
flask>=2.3.3
flask-cors>=4.0.0
pytz>=2024.1
